Split function package into separate packages for each dimension.
