+++ raysect/core/math/function/function1d/base.pxd
