+++ raysect/core/math/function/function2d/base.pyx
