+++ raysect/core/math/function/function3d/base.pyx
